changed auth system, added profile apis

+++ web/api/v1/auth_app/views.py
@@ -1,24 +1,21 @@
 import logging
-from django.dispatch import Signal
+
 from django.conf import settings
-from django.http import Http404
-from rest_framework import status
+from django.contrib.auth import get_user_model
 from django.utils.translation import gettext_lazy as _
-from rest_framework.exceptions import MethodNotAllowed
-from rest_framework.permissions import AllowAny, IsAuthenticated
-from rest_framework.generics import get_object_or_404, CreateAPIView, RetrieveAPIView
+from rest_framework.generics import CreateAPIView
+from rest_framework.generics import GenericAPIView
+from rest_framework.permissions import AllowAny
+from rest_framework.response import Response
 from rest_framework.status import HTTP_200_OK, HTTP_401_UNAUTHORIZED, HTTP_500_INTERNAL_SERVER_ERROR
 from rest_framework.views import APIView
 from rest_framework_simplejwt.exceptions import TokenError
 from rest_framework_simplejwt.tokens import RefreshToken
-from microservice_request.permissions import HasApiKeyOrIsAuthenticated
-from rest_framework_simplejwt.views import TokenVerifyView, TokenObtainPairView
+from rest_framework_simplejwt.views import TokenVerifyView, TokenRefreshView
+# from dj_rest_auth.views import LoginView
 
 from . import serializers
-from rest_framework.generics import GenericAPIView
-from django.contrib.auth import get_user_model
-from rest_framework.response import Response
-
+from .authentication_classes import AuthenticationByPhone
 from .services import AuthAppService
 
 User = get_user_model()
@@ -26,9 +23,73 @@ User = get_user_model()
 logger = logging.getLogger(__name__)
 
 
-class LoginView(GenericAPIView):
+class LoginEmailView(GenericAPIView):
     permission_classes = (AllowAny,)
-    serializer_class = serializers.LoginSerializer
+    serializer_class = serializers.LoginEmailSerializer
+
+    # def get_response(self):
+    #     serializer_class = self.get_response_serializer()
+    #
+    #     if getattr(settings, 'REST_USE_JWT', False):
+    #         from rest_framework_simplejwt.settings import (
+    #             api_settings as jwt_settings,
+    #         )
+    #         access_token_expiration = (timezone.now() + jwt_settings.ACCESS_TOKEN_LIFETIME)
+    #         refresh_token_expiration = (timezone.now() + jwt_settings.REFRESH_TOKEN_LIFETIME)
+    #         return_expiration_times = getattr(settings, 'JWT_AUTH_RETURN_EXPIRATION', False)
+    #
+    #         data = {
+    #             'user': self.user,
+    #             'access_token': self.access_token,
+    #             'refresh_token': self.refresh_token,
+    #         }
+    #
+    #         if return_expiration_times:
+    #             data['access_token_expiration'] = access_token_expiration
+    #             data['refresh_token_expiration'] = refresh_token_expiration
+    #
+    #         serializer = serializer_class(
+    #             instance=data,
+    #             context=self.get_serializer_context(),
+    #         )
+    #     elif self.token:
+    #         serializer = serializer_class(
+    #             instance=self.token,
+    #             context=self.get_serializer_context(),
+    #         )
+    #     else:
+    #         return Response(status=status.HTTP_204_NO_CONTENT)
+    #
+    #     response = Response(serializer.data, status=status.HTTP_200_OK)
+    #     if getattr(settings, 'REST_USE_JWT', False):
+    #         from .jwt_auth import set_jwt_cookies
+    #         set_jwt_cookies(response, self.access_token, self.refresh_token)
+    #     return response
+    #
+    # def login(self):
+    #     self.user = self.serializer.validated_data['user']
+    #     token_model = get_token_model()
+    #
+    #     if getattr(settings, 'REST_USE_JWT', False):
+    #         self.access_token, self.refresh_token = jwt_encode(self.user)
+    #     elif token_model:
+    #         self.token = create_token(token_model, self.user, self.serializer)
+    #
+    #     if getattr(settings, 'REST_SESSION_LOGIN', True):
+    #         self.process_login()
+
+    def post(self, request):
+        serializer = self.get_serializer(data=request.data)
+        serializer.is_valid(raise_exception=True)
+        user = serializer.user
+        handler = AuthAppService()
+        handler.login(user)
+        return Response(serializer.data)
+
+
+class LoginPhoneView(GenericAPIView):
+    permission_classes = (AllowAny,)
+    serializer_class = serializers.LoginPhoneSerializer
 
     def post(self, request):
         serializer = self.get_serializer(data=request.data)
@@ -37,26 +98,37 @@ class LoginView(GenericAPIView):
 
 
 class SignUpEmailView(CreateAPIView):
+    permission_classes = (AllowAny,)
     serializer_class = serializers.SignUpEmailSerializer
 
 
 class SignUpPhoneView(CreateAPIView):
+    permission_classes = (AllowAny,)
     serializer_class = serializers.SignUpPhoneSerializer
 
 
-class VerifyEmailView(CreateAPIView):
-    serializer_class = serializers.VerifyEmailSerializer
+class VerifyView(CreateAPIView):
+    permission_classes = (AllowAny,)
+    serializer_class = serializers.VerifySerializer
 
 
-class PasswordResetView(CreateAPIView):
-    serializer_class = serializers.PasswordResetSerializer
+class PasswordResetEmailView(CreateAPIView):
+    permission_classes = (AllowAny,)
+    serializer_class = serializers.PasswordResetEmailSerializer
+
+
+class PasswordResetPhoneView(CreateAPIView):
+    permission_classes = (AllowAny,)
+    serializer_class = serializers.PasswordResetPhoneSerializer
 
 
 class PasswordResetConfirmView(CreateAPIView):
+    permission_classes = (AllowAny,)
     serializer_class = serializers.PasswordResetConfirmSerializer
 
 
 class LogoutView(APIView):
+    permission_classes = (AllowAny,)
 
     def post(self, request, *args, **kwargs):
         response = self.full_logout(request)
@@ -67,6 +139,11 @@ class LogoutView(APIView):
             "detail": _("Successfully logged out.")
         },
             status=HTTP_200_OK)
+
+        if request.session.has_key('user_id'):
+            del request.session['user_id']
+            return response
+
         if cookie_name := getattr(settings, 'JWT_AUTH_COOKIE', None):
             response.delete_cookie(cookie_name)
         refresh_cookie_name = getattr(settings, 'JWT_AUTH_REFRESH_COOKIE', None)
@@ -95,4 +172,8 @@ class LogoutView(APIView):
 
 
 class VerifyJWTView(TokenVerifyView):
-    pass
+    permission_classes = (AllowAny,)
+
+
+class RefreshJWTView(TokenRefreshView):
+    permission_classes = (AllowAny,)
